refactor(settings): split account sign-in method rows into sibling module
